Added a new interface to handle converting read packets into requests whose responses are then fully parsed as HTTP responses.
The idea is that we need to keep track of both high level details so that we can properly reuse connections when appropriate and diagnose any issues due to replay itself.  The first implementation of this interface uses netty to setup a proxy to write one packet at a time to the target cluster and wraps the responses back.  For now, the wrapped response part isn't implemented.  Instead, when READ COMPLETE for the basic handler happens, we consider that done & emit a bogus summary.  Future edits will improve this a lot - but this commit just shows how to do basic proxying in netty.