Remove the proxy setup, replacing it w/ a simpler NettyPacketToHttpHandlerFactory, which only has an EventLoopGroup.
NettyPacketToHttpHandler stores the output channel and writes it directly from the source byte arrays from the replayer.  There wasn't a reason to write the data to a socket and run things through an input channel - UNLESS we wanted to get diagnostics on whether the HTTP requests were well-formed or not.  That can always be resurrected - look at commit a6151f12f89647b7ff0b1cebdb4694f5163c6b42 for the proxy code.

Moving away from the proxy, much of the code in FrontsideHandler has moved to NettyPacketToHttpHandler.  Now, there are a couple additional handlers in the (response) pipeline to parse the HTTP.  One of those handlers is HttpObjectAggregator, with a 1MB buffer.  That should really be knocked down to something that will stream the response bodies, but I haven't bothered to find such an class yet.  Response logging is split into logging individual packet responses (BacksideSnifferHandler) and confirming when the HTTP response has been completed (BacksideHttpWatcherHandler).  Both deal with an instance of AggregatedRawResponse to accumulate the response as it comes in and to send the built object to the callback once the response was done.  Still to be done are firing response callbacks when the HTTP connection wasn't cleanly terminated.

There's still lots of super sloppy logging and the code is still generally a mess.

Signed-off-by: Greg Schohn <schohn@amazon.com>