Added a main.py index checker tool
The tool currently uses the logstash config parser to retrieve source and target cluster configuration, then queries them both to compare indices and their configuration. It stops after printing a report of indices that already exist on the target cluster, either identical to the source cluster or with differing settings/mappings. Data for these indices cannot be safely migrated. It also prints the list of indices that can be safely created on the target cluster.

This commit also includes minor updates to logstash config parser, and an updated requirements.txt file

Signed-off-by: Kartik Ganesh <gkart@amazon.com>