Set up an image for ElasticSearch-7.10.2 with search-guard... Search-Guard is installed (from a remote zip plugin) and is configured with the plugin's 'install_demo_configuration' script. The elasticsearch.yml file is configured to support it coming up as a single node.
Additional changes are made to support our uses cases.  Two simple scripts are added to the container to toggle TLS on/off via the configuration.  TLS and HTTP auth are both disabled, as these are most useful for us at the moment.
Lastly, the image that is built from this docker file (which can also be done via gradle) is the base image for the capture proxy image.  That gets us the same TLS configurations and JDK.  We can choose then to run the logging proxy and Elasticsearch together in a single container, which already has all of the elasticsearch+search-guard artifacts, or to run them separately, using the base image for one container and the capture proxy image for the other.

To facilitate better dependency tracking of the 'external' docker images (those that aren't triggered by other gradle dependencies), I'm using a hash value of the Dockerfile's directory.  The dependent images now use the base image of a specific tag that matches that hash value.  That works well, but the docker compose file, nor anybody else, should ever have to know that tag, so two tags are created for each image, 'latest' and the hash value.

Signed-off-by: Greg Schohn <schohn@amazon.com>