Bugfixes for TrafficReplayer In NettyPacketToHttpHandler, consumeBytes() will no longer return null.  It will instead return a CompletableFuture that has completed exceptionally or has yet to complete.  There were some bugs around exceptions causing deadlock.  Within a ChannelFuture callback that was responsible for signaling the CompletableFuture that had been returned to the caller, close() or anything else that would throw (NPE) would cause the CompletableFuture never to complete, creating an eventual deadlock in the rest of the application.  Another issue, in consumeBytes was in a when consumeBytes was called recursively.  Since the CompletableFuture is of type Void, NULL is returned as the normal completion value.  However, the whenComplete incorrectly checked for non-null-ness for the value rather than a non-null exception.  That had a bonus hidden issue that since the exception was null, completedExceptionally() also threw an exception.  Now the branches have been swapped and deadlocks are gone.
For the BacksideHttpWatcherHandler, make sure that channelInactive() also triggers the callbacks.  The class will now set the aggregatedRawResponseBuilder to null to indicate that the callback HAS been called, so that it will not be called multiple times or that additional data won't be added to it AFTER the callback was triggered.  In the latter case, where data could hit afterward, addCallback will throw a RuntimeException.  Recall that BacksideSnifferHandler is responsible for adding components (packets) to the Builder.  This class should eventually take a Supplier that will simply return the built AggregratedRawResponse instead of an interface that's bigger than it needs.

ReplayEngine now has a close() method that is responsible for calling the callbacks for all of the remaining liveStreams.  TrafficReplayer, which provides those callbacks, will work through the two callbacks to corral all of the results that it can.

SourceTargetCaptureTuple had a simple race condition where the newline was sent in a separate outstream.write() from the contents.  Since the call isn't synchronized and is called from multiple threads, that would eventually be a problem.  The jsonFromHttpData() method also didn't deal with cases where the HttpMessage was malformed.  Now it will still create a JSONObject, but with and "Exception" field that includes the exception.toString() as the value.  I had tried to use message() as the value, but it isn't guaranteed to be non-null.

TrafficReplayer tries to do a better job of forwarding progress between the two callbacks and also keeping track of final residual work for every request/response.  That residual work (and the corresponding CompletableFutures) is combined together into a CompletableFuture that is used to trigger shutdown of the netty group.  That code also calls get(), so the stopGroup could also be after the get().

I've also attempted to calibrated log levels.

Signed-off-by: Greg Schohn <schohn@amazon.com>