Make testRoundTrip assertion more reliable (and a little looser)
The current assertion checks for a exact byte size, but this was failing in some cases due to an off-by-one error (677 actual bytes versus 676 expected bytes). Rather than simply update the exact byte value, I opted to make the check a little looser and incorporate a fudge factor of +/- 2 bytes.

Signed-off-by: Kartik Ganesh <gkart@amazon.com>