[index configuration tool] Change to parsing Data Prepper pipeline YAML as input (#215)
* Removal of Lark parsing

Signed-off-by: Kartik Ganesh <gkart@amazon.com>

* [index configuration tool] Change to parsing Data Prepper pipeline YAML as input

This change removes support for Logstash input files and the use of the Lark library. Instead, index_configuration_tool now accepts a Data Prepper pipeline YAML file as input.

Signed-off-by: Kartik Ganesh <gkart@amazon.com>

* [Index configuration tool] Added utility logic to test_constants.py

This allows easy updating of the stored pickle output file if/when the test pipeline YAML file is updated. This code is commented out for now since it will only be triggered manually.

Signed-off-by: Kartik Ganesh <gkart@amazon.com>

---------

Signed-off-by: Kartik Ganesh <gkart@amazon.com>