A critically important and longstanding bug has finally been fixed in NettyDecodedHttpRequestHandler.  When removing the HttpRequestDecoder after the HttpRequest (containing the headers, but not the payload) has been received, the HttpRequestDecoder DOES CORRECTLY send the remaining unparsed bytes through the pipeline.  For example, if a single packet contains the entirety of the message (headers + body) in a single ByteBuf, the HttpRequestDecoder's handleRemoved() DOES call fireChannelRead on the remaining bytes that had not yet been parsed.
HOWEVER, this wasn't noticed because the remaining ByteBuf was sent as a ByteBuf through the pipeline.  Two defects held the code back from handling this correctly.  The first was that the next handler was still the HttpDecodedHttpRequestHandler, because the pipeline was being cleared in ascending order (meaning request decoder was removed, THEN the decoded request handler).  Secondly, our HttpDecodedHttpRequestHandler didn't handle ByteBufs or any other messages nor did it pass them downstream.  It now will call super.channelRead() for all messages other than HttpContent and HttpRequest.  Those two changes have fixed the case where a single packet that needs its headers, but NOT its body, to be transformed to successfully have all of its contents sent through!

A number of handlers, expect the LastHttpContent message to hit them.  However, if the payload doesn't need to be transformed, we remove the HttpRequestDecoder, which is responsible for parsing the message and sending the LastHttpContent.    When LastHttpContent isn't received, NettySendByteBufsToPacketHandlerHandler will end up marking the request's transformation as an ERROR, even if the message headers were successfully transformed.  It's true that the HTTP message might have actually been incomplete, but from the HttpJsonTransformingConsumer and its pipeline's perspective, the transformation went exactly as expected, the message was sent out, and nothing else needs to be done.

To compensate for this gap, a new "EndOfInput" has been introduced that is sent as a sentinel when no more data will be added to the channel.  This is sent immediately before closing the channel down and only in cases where the HttpRequestDecoder has been removed, so that exactly one "End" message should be received by the handlers.  That's important so that handlers don't need to worry about tearing down or flushing resources and maintaining even more complicated state.

The above changes also put some more scrutiny on pipeline and handler hygiene, including the IncompleteJsonBodyException that is sent when there was a content error.  It now extends NoContentException() so that the top-level HttpJsonTransformingConsumer can simplify the handling of any kind of NoContentException with a redrive.  Any thrown IncompleteJsonBodyException itself is now handled in NettySendByteBufsToPacketHandlerHandler, which will pull the exception (and indeed any exception) forward into the currentFuture object that will be passed back to the HttpJsonTransformingConsumer after the pipeline has been closed.

The HttpJsonTransformingConsumer is now smarter about unwinding CompletionExceptions to get the inner cause, which it then determines if a redrive is appropriate or if a more serious and unknown exception has occurred.

TrafficReplayer has been refactored a bit and is connecting some diagnostic data a bit more, including Exception causes of errors.

I've also included a new bit of functionality to output request packets as a TrafficStream of reads that is compressed and base64 encoded.  My thought was to write unit tests to take those blocks in, but I haven't gotten around to wiring those up yet.  My thought is that we'll find strange patterns within our testing & being able to dump a short string that we can turn around and drop into a rig to reproduce will be valuable and a big effort saver.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>