Introduce a new class (OnlineRadixSorter) that takes a sequence of indexed objects and then visits the current contiguous monotonic sequence from the first item that had yet to be visited.  The javadoc has examples.  This is used because the TrafficStream sequences may arrive in the observed order.
The CapturedTrafficToHttpTransactionAccumulator now accepts potentially unordered TrafficStream objects by first passing those objects through an OnlnieRadixSorter, which maps the index values of the TrafficStream to index values for the sorter.  Once the TrafficStreams are in order, they’re sent to the existing, but slightly modified code to add observations to the accumulation.  Changes in the accumulation logic to support this include adding an OnlineRadixSorter to the Accumulation object.  This was done because it’s the value within the ExpiringTrafficStreamMap that the CapturedTrafficToHttpTransactionAccumulator uses.

However, the ExpiringTrafficStreamMap had coupled object addition (getOrCreate) with timestamp management and expiration.  Since a TrafficStream itself doesn’t have a timestamp (Observations within it do), the expiration logic has been pulled out of getOrCreate (which has been renamed getOrCreateWithoutExpiration).  The calling context is now responsible for calling expireOldEntries and get() still includes a timestamp argument that is updated each time an item is looked up.

Lastly, CapturedTrafficToHttpTransactionAccumulator was refactored so that the creation of an Accumulation was in the now outer-level accept() method for TrafficStreams rather than the inner-level one for Observations.  That necessitated changes in where edge conditions between writes and reads were handled.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>