Fix MIGRATIONS-1268. MIGRATIONS-1268 was fixed by starting the client channel up with a stock ChannelDuplexHandler that just forwards data to the next handler, which may or may not be an SslHandler.  The BacksideSnifferHandler is now installed AFTER the SslHandler if present, but still before all other handlers (other than the ChannelDuplexHandler used to initialize the channel).
The rest of the checkin is to build up out TLS testing facilities and testing facilities within the Replayer.  The code to randomly find a port to run an HTTP server on, and for the server itself, has been pulled from NettyScanningHttpProxyTest and placed into the new testUtilities project.  That code is now used by unit tests in both the trafficCaptureProxyServer and the trafficReplayer.   I've intentionally used client and server implementations different from netty since the real world will have clients that aren't just netty.  Lastly, AggregatedRawResponse now has a public helper function, getCopyOfPackets, that is more suitable for debugging and unit testing.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>