Add a cache class (ExpiringSubstitutableItemPool) that is to be used from within a netty event loop... which will eventually house client connections.
This cache allows callers to retrieve objects that were created within a window, while expiring objects that hadn't been pulled within that window.  Once objects are given to the caller (via getAvailableOrNewItem), they are removed from the cache.  Upon items leaving the cache, new items creation is begun to backfill the old items.  If there are no ready items to be retrieved, the oldest future to build an item is instead returned.  If there are no "in progress" items being built, then an item is supplied directly.  Creation order is maintained to try to minimize the perceived latency for callers.

The unit test for this class uses some concurrency primitives to mitigate the effects of time, but I'm not controlling the time used for expiration, so some of the expiration checks, which aren't as critical, could be unreliable when in a debugger.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>