Wire up the connection pool to the rest of the proxy.
After struggling with ChannelFuture being a Future<Void> extension and the ChannelFuture implementation itself having the final Channel value that we're interested in, it became apparent that the ExpiringSubstitutableItemPool needed to have more flexibility in storing and returning specific Future<> subclasses as determined by the context.  Doing that made it much simpler to get ChannelFutures back from the pool, which were MUCH more natural do deal with.

Other than that, some of the other work that happened here was wiring command line parameters.  A slightly larger bit of code deals with a couple extra levels of hierarchy between the BacksideConnectionPool and the ExpiringSubstitutableItemPool.  One level deals with different EventLoops (threads) via a FastThreadLocal (netty's version of ThreadLocal) and the second is a HashMap that organizes different pools for each of those threads by channel class, since we need to make sure that we're returning the right kinds of channels.

Lastly, I'm not sure that the SSL handshaking to the destination server would always completely finish before we tried to write (or read) data on the channel.  This was an issue for the Replayer (see https://github.com/opensearch-project/opensearch-migrations/pull/238).  I think that the solution here, where a new ChannelFuture is bound to the SSLHandler's handshakeFuture() is a much cleaner approach.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>