Further changes to unify the replayer's auth story and to tie sigv4 signing all the way back to the command line.
The command line now takes a service and region as opposed to just a zero arity boolean flag as both are required for SigV4 calculations.  These values should refer to the region and service that the replayer is making calls to (e.g., the 'target' server).
Notice that an AuthTransformerFactory is passed around that returns an AuthTransformer for a specific http message.
There could be different strategies dependent upon the original message.  For example, unauthorized users might simply be mapped to no transformation, meaning that we don't need to go through the added effort of buffering and signing the messages.
The old style http header transformation via jolt is also gone and has been replaced with an (HeaderOnly)AuthTransformer class that manipulates headers.  I'm not happy with the care that I had to put into the  NettyDecodedHttpRequestHandler class (now NettyDecodedHttpRequestPreliminaryConvertHandler) to do the auth transformation in two spots while also passing more complicated auth transformations into the pipeline construction helper methods.
I've also converted the SigV4Signer to be an AuthTransformer and have tied it into the netty handler to host it, though none of that has been tested in this commit.

Unrelated changes include converting the packets of an HttpMessage(AndTimestamp) from an ArrayList to a named extension (ie. typedef).  I've never been comfortable with the vagueness of the ArrayList.  I had also been working on edits to take the response into the AuthTransformer so that 403s could be ignored - but decided that would be a bad idea.  1) Not every request will have a response and 2) it will allow future requests that should have been blocked to never be tested and the time to test them will be within the target server.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>