Cleanup and bugfixes to get SigV4 more correctly wired with the rest of the replayer.
I've got a unit test to verify that the contents of the signature match a mock key and mock date.  To put that together, I've extracted another helper function from a similar test and placed it into TestUtils.
There was one other bugfix to remove a spurious "toString()" that followed a headers.put call for the content-length.  The string wasn't used and put() will return null when the header wasn't previously present, creating an NPE situation.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>