MIGRATIONS-1250.  Setup NettyPacketToHttpConsumer to reuse channels... and to cache them so that they're associated with the original connection ids.  That enables us to honor keep-alived connections when doing replays.
To support that, a lot of cleanup and some bugfixes had to happen.  The NettyPacketToHttpConsumer initiates the SslHandler and handles the handshake similarly to how the capture proxy does it.  Also, like the capture proxy, autoread is now set on the channel once the handlers are in place.  finalizeRequest() is responsible for returning a CompletableFuture that will tear down the pipeline as its last (asynchronous) step so that when the channel is passed back into a new NettyPacketToHttpConsumer, the pipeline can be reinitialized (for everything post SslHandler, if applicable).

I've also added a couple helper functions to Utils to support converting a sequence of packets to a string, either by printing them out within brackets ("[]") or by pumping them through an EmbeddedChannel with Http decoders.  There's a helper function and a ThreadLocal to manage this more easily for users using toString(), etc that wouldn't normally be able to pass flags into functions.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>