Split the PacketToTransformingHttpHandlerFactory into two factories.
A new "PacketToTransmitFactory" now houses all of the setup for the NettyPacketToHttpConsumer as well as the cache and logic to kill all of the active connections.  TrafficReplayer now maintains TWO different factories.  The first produces HttpJsonTransformingConsumers which convert the original request packets into a new stream (with the help of a new packetReceiver that simply stores ByteBufs into an ArrayList).  Those output lists are then sent by the TrafficReplayer to the netty receiver.  Breaking up the responsibility to transform and send packets at a high level allows the code to more easily plan for when when packets are ready to send and should create reduce deviation in times that requests are sent, at least due to computing new messages.

I've also refactored some of the simplest POJO structures into a datatypes package.  When I pulled the HttpRequestTransformationStatus into its own class, it was too much namespace pollution at the top level.  The whole repo is in bad need of reorganization, but I was beginning to lose track of some of the really significant classes.

Some care should be taken to confirm that the original transformation pipeline and all of its cached objects have been purged/are eligible for GC, but that check hasn't happened yet.  The next phase will further separate the netty sending receiver from transformation by associating those sends with timed queues.  Meanwhile, on the transformation side, the producer feeding into it still needs to implement backpressure, or a pull model rather than a push one.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>