A number of simplifications to reduce baggage on various types to eliminates the need to generify more classes.
For example, AggregatedRawResponse no longer returns the request packets.  That strips the class down to JUST stuff related to the response.  If a caller doesn't care about anything but the response, this is _still_ the class for them.  Request packets are now pulled into the TransformedTargetRequestAndResponse (former AggregatedTransformedResponse) subclass.  That gives the entire context for a single request.  During that refactoring, I noticed that the recent change to add an additional errorCause to the AggregatedTransformedResponse was unnecessary since the base class already had it.  The transformed request itself needed some rework since it was dealing with ByteBufs.  It no longer directly extends a List so that it can return a stream where each ByteBuf object is duplicated by the returned stream so that callers each get their own isolated access to read indices.

This change continues to migrate the responsibility of sending requests into the RequestSenderOrchestrator, which has required opening up access to fields around other classes (which could be more graceful).  However, there aren't yet controls to schedule when requests are sent, so exceptions will be thrown when connections are reused, which I'll be addressing shortly.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>