Bugfixes, thanks to the OnlineRadixSorter, to get the unit test to pass.
There was a race condition in shuffling work onto the EventLoop thread.  Since UniqueRequestKeys are being passed around, we have access to the total ordering of all requests, even on the other side of the race condition.  This change takes advantage of that by adding an OnlineRadixSorter and processing all schedules first through that sorter, which is sorting on the request component of the UniqueRequestKey.  That means that each schedule is happening in chronological order w/out the need to worry about gaps.

In addition, the unit test has been beefed up (as well as the code) to schedule some request sends with the same start timestamp.  That required the addition of a deque to make the TimeToResponseFulfillmentFutureMap something like a stable multimap.  The test also now breaks the sends up into separate characters to try to stress the system a bit more.

Lastly, I've reduced logging in general, pushing most things down to trace and leaving things that were more truly edge cases at info.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>