Lay foundations to add support for TrafficCaptureSources to block when too many TrafficStream objects are still in flight throughout the rest of the replayer.
The ITrafficCaptureSource itself has changed from being a blocking interface to a non-blocking one that returns a CompletableFuture with a List<TrafficStream> or completes exceptionally with an EOF if we've reached EOF (which isn't currently possible with Kafka).  The way that this was implemented with blocking calls made it exceptionally difficult to write unit tests (which are still in-progress).

The InputStreamOfTraffic and KafkaProtobufConsumer keep most of their runtime behavior, though their structure is transposed to return values rather than send them through a consumer.  Both sources use CompletableFuture.supplyAsync since I didn't immediately know of a way to get protobuf's mergeDelimitedFrom or Kafka to work directly asynchronously (I'm sure that the latter could be done much better).  The CompletableFuture returned is a List<TrafficStream> because that's an API closer to what Kafka's source can provide.  Rather than create a separate CompletableFuture for each component for the case that will be used at scale, I chose to optimize for the Kafka scenario - which also didn't require me to do more complicated buffering within that class or elsewhere.

I've also done some refactoring between the TrafficReplayer and ReplayEngine to make the ReplayEngine deal with the TrafficSource (which it interfaces through its BufferedTimeController interface) as it processes the stream's scheduling calls.  Time shifting has moved from the RequestSenderOrchestrator to the ReplayEngine because the former has enough going on and I found a bug where I was only adjusting the start times for packet sends, but not the intervals.  There were too many one-off time changes in a class with other responsibilities, so I've extracted the calculations into their own class and now have ReplayEngine drive those transformations.  The TimeShifter itself needed an inverse so that we can take a real time that we've completed a scheduled event and convert it back into a source time so that we can adjust the buffer window.  That's critical since we might not get any traffic for a period of time greater than our buffer window, which would otherwise cause the source to stall w/out periodic updates from the ReplayEngine.

Still to be addressed:

The ReplayEngine needs to wire up the periodic updates of the source time, even if there wasn't traffic.

This change should catastrophically truncate replays since there are currently no calls to advanceReadHorizonTo(), though that's coming shortly.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>