Fix the BlockingTrafficSource by adding a while() loop instead of an if to re-check the blocking condition to determine if the code should proceed. That fixes some race conditions where the signal for the last block was fired after the last caller had gotten through, but because the time check passed for it and there wasn't a reason for it to block.  That left a block up for grabs, which was immediately available since the semaphore had a permit.  The fix is to simply get the signal, but recheck, waiting on the signal again.  In rare cases there's an oversignaling here, but I think its unavoidable, or at least not worth worrying about at the moment.
I'm purposefully leaving some sleep statements within the code that I had used to fuzz the test runs.  Notice that they'll be removed next commit.  Ideally, this would be managed by passing an executor in that the class uses for its supplyAsync operations and could be injected by test code.  That's also work for later though.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>