[Fetch Migration] Add support for inline pipeline via environment variable
The orchestrator now checks for an INLINE_PIPELINE environment variable. If present, its value is then written out as the Data Prepper pipeline config to be used. This value is expected to be a base64-encoded string. This feature removes the need for the wrapper Docker image - the CDK now uses the FetchMigration Docker image directly. Finally, the CDK uploads the inline pipeline string to AWS Secrets Manager so it can be imported (via ARN) into the ECS Container when the Fetch Migration task is run.

Signed-off-by: Kartik Ganesh <gkart@amazon.com>