More memory leak work including fixes to get all unit tests to pass with no leaks.
Fixed a handful of over-retains within the handlers.  I also did general cleanup to remove dead code and problematic test code.

The WrapWithNettyLeakDetection annotation now has a boolean to disable leak checks.  It also now can apply to all of the tests within a class (with the method annotation overriding the class one if both are present).  That's useful to only run the test once and to document that we don't believe that we need those tests.  I'd like to refactor our entire codebase into packages that show different kinds of entanglements so that these relationships will be easier to audit in the future.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>