Checkpoint - Fix some bugs and improve some log messages. This code fixes a major bug that was triggered when a TrafficStream's first set of records was missing its first (0) index.  That caused the BlockingTrafficSource to wait for the first TrafficStream to complete, but it never would.  How we manage bumping time at the beginning of the application has been significantly rethought.  We should still test that expirations are carrying through to bump all of the other cases.
The logging messages for when the expiration timeout window weren't big enough weren't helpful.  I've reworked those to provide the correct values and to provide stronger guidance to the user.

Other refactoring in a number of spots to try to improve out logging stories, including making the coreUtilities project dependent upon the captureProtobuf project so that it can add some common, handy code.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>