WIP: Work on limiting the number of TrafficStreams/requests that can be present at any time.
Most of the lines changed are due to making TrafficStreamKeys into an interface so that I can wrap them and extend them.

Changes that were more significant include adding a couple fields of metadata to each TrafficStream object.  I haven't modified the capture code yet to fill those values in, but I do have a class in the replayer test directory that converts older versions of streams into the newer version by augmenting the TrafficStreams with metadata that was pulled in from previous TrafficStreams.

The reason that metadata is now required is because every TrafficStream object needs to stand on its own.  As the codebase was, and still is as of this commit, TrafficStreams are resequenced and ALL indices for a given connection must be present before any of the requests from the stream are accumulated and replayed.  Going forward, we want/need to commit records from Kafka on a per-record basis.  Many different connections will have TrafficStreams jumbled together, so the last TrafficStream for one connection could be surrounding many other TrafficStream sequences.  That means that we'll be issuing commits to Kafka at many partial points of connections' streams.  Notice that we won't be able to commit past a point where there could be some data that we still need to send/process.  However, we WILL run into many situations where there will be ambiguity at the first reads of a TrafficStream with index > 0.  When the prior TrafficStream isn't available, it isn't possible to determine when the first observations were reads if they were for a new request or one continued from the last TrafficStream.  The new metadata field for lastObservationWasRequestRead give the downstream processors enough information to reconstruct the requests within this TrafficStream and beyond.  requestCount lets us apply a sequential ordering to all of the requests within a TrafficStream.

BlockingTrafficStream has added the ability to track costs for each TrafficStream that has been returned by its underlying source.  Those cost values are then used as the argument to acquire for a semaphore.  The caller constructs BlockingTrafficStream with a maximum cost that is permissible and two lambda functions.  One function computes the cost, which is calculated before the TrafficStream object is returned by readNextChunk() and is persisted by the BlockingTrafficStream.  The other lambda calculates how many requests are within the TrafficStream so that the BlockingTrafficSource can countdown the number of times that each TrafficStream has a request that is marked as done.  Once that counter reaches 0, the cost is released from the semaphore and the tracking item is removed.

Work is still remaining to process TrafficStreams immediately, rather than waiting for them within the OnlineRadixSorter.  Changes need to be made to the capture kafka code to enforce relative stream ordering so that we don't need to do sorting.  Strict ordering is a critical requirement now so that we don't need to look ahead for prior TrafficStream objects.  That work will also remedy an issue that currently is causing the replayer to deadlock when TrafficStream objects are pending due to earlier TrafficStreams not being available.  However, we'll never even check for later ones or advance the cursor to expire elements because we're stuck waiting because the cost semaphore is waiting for the completion of those waiting TrafficStreams.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>