[Fetch Migration] Adding orchestration script for Fetch Migration steps (#294)
This change adds an "orchestration" Python script that runs the Fetch Migration steps in order:

1) The metadata migration script is run first to compare the indices on source and target clusters. A human-readable report is printed, and an updated Data Prepper pipeline YAML file is written to the Data Prepper `pipelines` directory

2)  Next, Data Prepper is kicked off as a [sub-process](https://docs.python.org/3/library/subprocess.html) to migrate data based on the output pipeline YAML file

3) After the Data Prepper process is kicked off, the monitoring script is run to poll DP's Prometheus metrics endpoint and determine when migration is complete and the process can be shut down

The orchestrator script takes multiple command-line inputs to enable execution of these steps. The Dockerfile definition has also been updated to use the orchestrator script as the entrypoint.

* Adding logging to migration_monitor.py

Add logging to track progress of the migration. This commit also moves the sleep state to the start of the loop. This allows an initial 30 seconds to go by, letting the Data Prepper process startup. Otherwise, we run the risk of the GET call failing.

---------

Signed-off-by: Kartik Ganesh <gkart@amazon.com>