All 293 test cases for ExhaustiveCapturedTrafficToHttpTransactionAccumulatorTest now pass.
The original TrafficStreamKeys are now used to create each new transaction ("rrPair").  That change was critical for two reasons.  It's going to allow the code to send back the same key that the TrafficCaptureSource provided so that internal information (like Kafka commit sync point) can be leveraged.  It also required rotating the code around so that a new accumulating transaction wasn't spun up as soon as the old one was finished, but when the new one started.  The trafficKey and the **original** source offset needed to be pulled back in for each of the UniqueRequestIds assigned for each transaction.  To finish off making sure that the right TrafficStreamKeys are being held, the Close observation's handler calls hold() on the key before invoking the callback.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>