[Fetch Migration] Handling of idle Data Prepper pipeline with incomplete migration
This includes a new ProgressMetrics class that is used by the migration monitor to track various Data Prepper and API failure metrics in order to detect an idle pipeline. Much of the migration-success logic from the monitoring module has now been encapsulated in this class. Unit test updates and improvements are also included.

Signed-off-by: Kartik Ganesh <gkart@amazon.com>