Continue work on making the replayer more reliable to the point where each TrafficStream is being committed after it's been fully processed. Along the way, partially to build up the tools for testing that behavior, partially because other bugs were discovered along the way, a number of bugfixes, enhancements, and refactorings continue to pile up.
Accumulations now track the TrafficKey instead of the RequestResponsePair.  There are moments (since the past couple commits) where the rrPair field may be null and we might not have access to the trafficStreamKey.  This change was made a bit easier since I've also refactored the CapturedTrafficToHttpTransactionAccumulator to take in a help object of callbacks rather than 3 callbacks.  It's actually now 4 callbacks because of a new one that will be required for commits of TrafficStreams that are expired that won't be flushed.  Accumulations no longer reset the last timestamp that came in.  Doing that means that we can use the same value within fireAccumulationsCallbacksAndClose, which was using a value from rrPair, but as mentioned, that could be null under the newer processing model.

In addition to that, there's a notion of shutdown within the TrafficReplayer.  A number of additional Atomic fields were added.  These are most valuable at the moment from test code so that I can tear a TrafficReplayer down and bring a subsequent one up with a totally fresh state EXCEPT for my tupleReceiver callback (which has also been newly refactored).  That structure will allow me to verify that the commit points that we've reached were accurate and that the flow of "at least once" message delivery is present for all requests.  That last part isn't being verified yet, but the test should complete after shutting down a series of TrafficReplayers.  There ARE assertions in place to make sure that old replayer signals (calls to the callback) aren't happening after we've torn down an old TrafficReplayer and created a new one.

To support shutdowns, a number of minor changes were made.  The policy is to basically shed work as fast/abruptly as possible, with the idea being that we'll just rerun requests a second time through.  Also to support the new unit test, Error (as in Throwable) handling is cleaner now.  OOMErrors and other Errors should now trigger a shutdown of the replayer and cause a quick exit.  This is being tested by the FullReplayerTest as it shuts down TrafficReplayers by throwing an Error, so I get to test two features in one test.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>